Refactor Rename auth into helper
Dringendes Umbennenen notwendig da in diesem Ordner alle Helper Dateien und funktionen liegen anstelle von nur Authentifizierung
