+++ server/helper/authenticate.js
