+++ server/routes/contactlist.js
@@ -1,6 +1,6 @@
 const express = require('express');
 const router = express.Router();
-const util = require('../auth/utilities');
+const util = require('../helper/utilities');
 
 router.post('/', (req, res, next) => {
     req.models.Contactlist.create({user_id1: req.userid, user_id2: req.user_id, status: 0}).then((result) => {
